+++ packages/alpha/src/__tests__/api/simulationRunsApi.int.spec.ts
@@ -1,6 +1,7 @@
 // Copyright 2023 Cognite AS
 
 import { describe, expect, test } from 'vitest';
+import type CogniteClientAlpha from '../../cogniteClient';
 import { setupLoggedInClient } from '../testUtils';
 import {
   fileExtensionTypes,
@@ -68,7 +69,6 @@ describeIf('simulator runs api', () => {
         name: 'Test Simulator Model',
         description: 'Test Simulator Model Desc',
         dataSetId: 97552494921583,
-        labels: [{ externalId: 'air-quality-po-1' }],
         type: 'WaterWell',
       },
     ]);