+++ packages/alpha/src/__tests__/api/simulatorModelsApi.spec.ts
@@ -1,6 +1,7 @@
 // Copyright 2023 Cognite AS
 
 import { describe, expect, test } from 'vitest';
+import type CogniteClientAlpha from '../../cogniteClient';
 import { setupLoggedInClient } from '../testUtils';
 import {
   fileExtensionTypes,
@@ -46,12 +47,22 @@ describeIf('simulator models api', () => {
         name: 'Test Simulator Model',
         description: 'Test Simulator Model Desc',
         dataSetId: 97552494921583,
-        labels: [{ externalId: 'air-quality-po-1' }],
         type: 'WaterWell',
       },
     ]);
     expect(res.length).toBe(1);
     expect(res[0].externalId).toBe(modelExternalId);
+  });
+
+  test('aggregate models', async () => {
+    const aggregateResponse = await client.simulators.aggregateModels({
+      filter: {
+        simulatorExternalIds: [simulatorExternalId],
+      },
+      aggregate: 'count',
+    });
+
+    expect(aggregateResponse[0].count).toBeGreaterThan(0);
   });
 
   test('list models', async () => {