+++ packages/alpha/src/api/simulators/routinesApi.ts
@@ -1,11 +1,6 @@
 // Copyright 2024 Cognite AS
 
-import {
-  BaseResourceAPI,
-  type CDFHttpClient,
-  type IdEither,
-  type MetadataMap,
-} from '@cognite/sdk-core';
+import { BaseResourceAPI, type IdEither } from '@cognite/sdk-core';
 import type {
   SimulatorRoutine,
   SimulatorRoutineCreate,
@@ -20,18 +15,18 @@ export class RoutinesAPI extends BaseResourceAPI<SimulatorRoutine> {
     return this.pickDateProps(['items'], ['createdTime', 'lastUpdatedTime']);
   }
 
-  public create = async (items: SimulatorRoutineCreate[]) => {
+  public create = (items: SimulatorRoutineCreate[]) => {
     return this.createEndpoint(items);
   };
 
-  public list = async (filter?: SimulatorRoutineFilterQuery) => {
+  public list = (filter?: SimulatorRoutineFilterQuery) => {
     return this.listEndpoint<SimulatorRoutineFilterQuery>(
       this.callListEndpointWithPost,
       filter
     );
   };
 
-  public delete = async (ids: IdEither[]) => {
+  public delete = (ids: IdEither[]) => {
     return this.deleteEndpoint(ids);
   };
 }