+++ packages/alpha/src/api/simulators/routinesRevisionsAPI.ts
@@ -2,8 +2,12 @@
 
 import { BaseResourceAPI, type IdEither } from '@cognite/sdk-core';
 import type {
+  CursorResponse,
+  HttpResponse,
   SimulatorRoutineRevision,
+  SimulatorRoutineRevisionBase,
   SimulatorRoutineRevisionCreate,
+  SimulatorRoutineRevisionView,
   SimulatorRoutineRevisionsFilterQuery,
 } from '../../types';
 
@@ -12,25 +16,44 @@ export class RoutineRevisionsAPI extends BaseResourceAPI<SimulatorRoutineRevisio
    * @hidden
    */
   protected getDateProps() {
-    return this.pickDateProps(['items'], ['createdTime', 'lastUpdatedTime']);
+    return this.pickDateProps(['items'], ['createdTime']);
   }
 
-  public create = async (items: SimulatorRoutineRevisionCreate[]) => {
+  public create = (items: SimulatorRoutineRevisionCreate[]) => {
     return this.createEndpoint(items);
   };
 
-  public list = async (filter?: SimulatorRoutineRevisionsFilterQuery) => {
-    return this.listEndpoint<SimulatorRoutineRevisionsFilterQuery>(
-      this.callListEndpointWithPost,
-      filter
+  private callRevisionListEndpoint = async <
+    RevisionResponseType extends SimulatorRoutineRevisionBase,
+  >(
+    query?: SimulatorRoutineRevisionsFilterQuery
+  ): Promise<HttpResponse<CursorResponse<RevisionResponseType[]>>> => {
+    const response = await this.post<CursorResponse<RevisionResponseType[]>>(
+      this.listPostUrl,
+      {
+        data: query || {},
+      }
     );
+    return response;
+  };
+
+  public list = <
+    RevisionResponseType extends
+      SimulatorRoutineRevisionBase = SimulatorRoutineRevisionView,
+  >(
+    query?: SimulatorRoutineRevisionsFilterQuery
+  ) => {
+    return this.cursorBasedEndpoint<
+      SimulatorRoutineRevisionsFilterQuery,
+      RevisionResponseType
+    >(this.callRevisionListEndpoint, query);
   };
 
   public retrieve(items: IdEither[]) {
     return this.retrieveEndpoint(items);
   }
 
-  public delete = async (ids: IdEither[]) => {
+  public delete = (ids: IdEither[]) => {
     return this.deleteEndpoint(ids);
   };
 }