+++ packages/stable/src/__tests__/api/files.unit.spec.ts
@@ -2,8 +2,7 @@
 
 import nock from 'nock';
 import { beforeEach, describe, expect, test } from 'vitest';
-import { setupMockableClient } from '../testUtils';
-import { mockBaseUrl } from '../testUtils';
+import { mockBaseUrl, setupMockableClient } from '../testUtils';
 
 describe('Files unit test', () => {
   let client: CogniteClient;