+++ packages/stable/src/__tests__/api/filesMultiPart.unit.spec.ts
@@ -1,10 +1,12 @@
+// Copyright 2020 Cognite AS
+
 import { sleepPromise } from '@cognite/sdk-core';
 import { mockBaseUrl } from '@cognite/sdk-core/src/__tests__/testUtils';
 import nock from 'nock';
 import { beforeEach, describe, expect, it, test } from 'vitest';
 import type { FilesMultipartUploadSessionAPI } from '../../api/files/filesMultipartUploadSessionApi';
 import type { MultiPartFileChunkResponse } from '../../types';
-import { setupMockableClientForUnitTest } from '../testUtils';
+import { setupMockableClient } from '../testUtils';
 
 describe('Multi part upload unit test', () => {
   let client: CogniteClient;
@@ -27,7 +29,7 @@ describe('Multi part upload unit test', () => {
     };
   }
   beforeEach(() => {
-    client = setupMockableClientForUnitTest();
+    client = setupMockableClient();
     nock.cleanAll();
   });
 
@@ -64,7 +66,7 @@ describe('Multi part upload unit test', () => {
           numberOfParts,
           true
         )
-      ).rejects.toThrowError('parts must be greater than 0 and less than 250');
+      ).rejects.toThrowError('parts must be in range 1 <= parts <= 250');
       expect(initAPiNock.isDone()).toBeFalsy();
     }
   );