+++ packages/stable/src/api/files/filesApi.ts
@@ -2,8 +2,10 @@
 
 import {
   BaseResourceAPI,
+  type CDFHttpClient,
   type CursorAndAsyncIterator,
   type HttpHeaders,
+  type MetadataMap,
   sleepPromise,
 } from '@cognite/sdk-core';
 import type {
@@ -21,9 +23,22 @@ import type {
   IdEither,
   IgnoreUnknownIds,
   ItemsWrapper,
+  MultiPartFileUploadResponse,
 } from '../../types';
+import { FilesMultipartUploadSessionAPI } from './filesMultipartUploadSessionApi';
 
 export class FilesAPI extends BaseResourceAPI<FileInfo> {
+  private limits = {
+    minimumNumberOfParts: 1,
+    maxNumberOfParts: 250,
+  };
+  private _client: CDFHttpClient;
+  private _map: MetadataMap;
+  private _baseUrl: string;
+  constructor(...args: [string, CDFHttpClient, MetadataMap]) {
+    super(...args);
+    [this._baseUrl, this._client, this._map] = args;
+  }
   /**
    * Specify that dates should be parsed in requests and responses
    * @hidden
@@ -232,6 +247,55 @@ export class FilesAPI extends BaseResourceAPI<FileInfo> {
       { data: { items } }
     );
     return this.addToMapAndReturn(response.data.items, response);
+  }
+  /**
+   * [Init a multipart file upload](https://api-docs.cognite.com/20230101/tag/Files/operation/initMultiPartUpload)
+   *
+   * ```js
+   * // Each file part must be larger than 5 MiB, and smaller than 4000 MiB. The file part for the last uploadURL can be smaller than 5 MiB.
+   * const numberOfParts = 5;
+   * // automatic upload:
+   * const multiPartUploadApi = await client.files.multipartUploadSession({name: 'examplefile.jpg', mimeType: 'image/jpeg'}, numberOfParts);
+   * ```
+   */
+  public async multipartUploadSession(
+    fileInfo: ExternalFileInfo,
+    parts: number,
+    overwrite = false
+  ) {
+    const response = await this.getMultipartUploadSession(
+      fileInfo,
+      parts,
+      overwrite
+    );
+    const multipartUploadSession = new FilesMultipartUploadSessionAPI(
+      this._baseUrl,
+      this._client,
+      this._map,
+      response.data
+    );
+    return this.addToMapAndReturn(multipartUploadSession, response);
+  }
+  private async getMultipartUploadSession(
+    fileInfo: ExternalFileInfo,
+    parts: number,
+    overwrite = false
+  ) {
+    if (
+      parts < this.limits.minimumNumberOfParts ||
+      parts > this.limits.maxNumberOfParts
+    ) {
+      throw Error(
+        `parts must be in range ${this.limits.minimumNumberOfParts} <= parts <= ${this.limits.maxNumberOfParts}`
+      );
+    }
+    const path = this.url('initmultipartupload');
+    const params = { overwrite: overwrite, parts: parts };
+    const response = await this.post<MultiPartFileUploadResponse>(path, {
+      params,
+      data: fileInfo,
+    });
+    return response;
   }
 }