+++ packages/stable/src/types.ts
@@ -3432,6 +3432,15 @@ export interface UnitSystem {
   name: string;
   quantities: UnitSystemQuantity[];
 }
+export interface MultiPartFileUploadResponse extends FileInfo {
+  uploadUrls: string[];
+  uploadId: string;
+}
+
+export interface MultiPartFileChunkResponse {
+  partNumber: number;
+  status: number;
+}
 
 export {
   AggregatedHistogramValue,
@@ -3571,3 +3580,5 @@ export {
   ViewDefinitionProperty,
   ViewPropertyDefinition,
 } from './api/models/types.gen';
+
+export * from './api/sessions/types';