+++ packages/alpha/CHANGELOG.md
@@ -3,6 +3,85 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.35.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.35.1...@cognite/sdk-alpha@0.35.2) (2025-01-17)
+
+
+### Bug Fixes
+
+* **simint:** add simulation run example ([#1188](https://github.com/cognitedata/cognite-sdk-js/issues/1188)) ([a912037](https://github.com/cognitedata/cognite-sdk-js/commit/a912037c1c4afaddffef6d644decc4ec561b6def))
+
+
+
+
+
+## [0.35.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.35.0...@cognite/sdk-alpha@0.35.1) (2024-12-18)
+
+
+### Bug Fixes
+
+* use correct type for aggregate query ([#1180](https://github.com/cognitedata/cognite-sdk-js/issues/1180)) ([f2e30c6](https://github.com/cognitedata/cognite-sdk-js/commit/f2e30c653324d9c991afa3c981494f33e4a42533))
+
+
+
+
+
+# [0.35.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.34.1...@cognite/sdk-alpha@0.35.0) (2024-12-18)
+
+
+### Features
+
+* add aggregate to modelsApi ([#1179](https://github.com/cognitedata/cognite-sdk-js/issues/1179)) ([7173057](https://github.com/cognitedata/cognite-sdk-js/commit/7173057b76e4ec6a7dc66a226eea864ba08e89f8))
+
+
+
+
+
+## [0.34.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.34.0...@cognite/sdk-alpha@0.34.1) (2024-12-12)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
+# [0.34.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.33.0...@cognite/sdk-alpha@0.34.0) (2024-12-05)
+
+
+### Features
+
+* use enum for connector status ([#1175](https://github.com/cognitedata/cognite-sdk-js/issues/1175)) ([ad0d6d9](https://github.com/cognitedata/cognite-sdk-js/commit/ad0d6d9c3243cd1121d1c02f462e552107319fd4))
+
+
+
+
+
+# [0.33.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.32.2...@cognite/sdk-alpha@0.33.0) (2024-11-21)
+
+
+### Features
+
+* **alpha:** multiple updates for simint APIs ([#1174](https://github.com/cognitedata/cognite-sdk-js/issues/1174)) ([96bddf1](https://github.com/cognitedata/cognite-sdk-js/commit/96bddf183eb46c3edb97be793edb8e8e8e5e857f))
+
+
+
+
+
+## [0.32.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.32.1...@cognite/sdk-alpha@0.32.2) (2024-11-07)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
+## [0.32.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.32.0...@cognite/sdk-alpha@0.32.1) (2024-10-07)
+
+**Note:** Version bump only for package @cognite/sdk-alpha
+
+
+
+
+
 # [0.32.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-alpha@0.31.3...@cognite/sdk-alpha@0.32.0) (2024-09-17)