+++ packages/beta/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.12.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.12.2...@cognite/sdk-beta@5.12.3) (2025-01-17)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.12.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.12.1...@cognite/sdk-beta@5.12.2) (2024-12-12)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+## [5.12.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.12.0...@cognite/sdk-beta@5.12.1) (2024-11-07)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+
+
+
+
+# [5.12.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.11.6...@cognite/sdk-beta@5.12.0) (2024-10-07)
+
+
+### Features
+
+* **files:** move multipart functionality to stable [BND3D-4545] ([#1132](https://github.com/cognitedata/cognite-sdk-js/issues/1132)) ([1e33003](https://github.com/cognitedata/cognite-sdk-js/commit/1e33003f364ad0932cca24ea2a9d457ac382f9ec))
+
+
+
+
+
 ## [5.11.6](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.11.4...@cognite/sdk-beta@5.11.6) (2024-08-26)