+++ packages/stable/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.17.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@9.17.0...@cognite/sdk@9.17.1) (2025-01-17)
+
+**Note:** Version bump only for package @cognite/sdk
+
+
+
+
+
+# [9.17.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@9.16.1...@cognite/sdk@9.17.0) (2024-12-12)
+
+
+### Features
+
+* **stable:** add support for sessions API ([#1177](https://github.com/cognitedata/cognite-sdk-js/issues/1177)) ([28b0fdd](https://github.com/cognitedata/cognite-sdk-js/commit/28b0fddf880091508238fea416b87d72cea32529))
+
+
+
+
+
+## [9.16.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@9.16.0...@cognite/sdk@9.16.1) (2024-11-07)
+
+
+### Bug Fixes
+
+* **files:** arrange-imports for multipart session ([#1171](https://github.com/cognitedata/cognite-sdk-js/issues/1171)) ([0848b3f](https://github.com/cognitedata/cognite-sdk-js/commit/0848b3f57bf788cee8d991b76b19a647590a063d))
+
+
+
+
+
+# [9.16.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@9.15.7...@cognite/sdk@9.16.0) (2024-10-07)
+
+
+### Features
+
+* **files:** move multipart functionality to stable [BND3D-4545] ([#1132](https://github.com/cognitedata/cognite-sdk-js/issues/1132)) ([1e33003](https://github.com/cognitedata/cognite-sdk-js/commit/1e33003f364ad0932cca24ea2a9d457ac382f9ec))
+
+
+
+
+
 ## [9.15.7](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk@9.15.5...@cognite/sdk@9.15.7) (2024-08-26)